Initial Commit
This set of instruments is intended to help controller manufacturers (and others) test if the gyroscope is set up for "Real World Sensitivity" - i.e. 1 rotation in the real world = 1 rotation according to accumulated sensor readings.
It is expected that integrating a gyro packet by the time since last packet should show the real world rotation since last packet polled. To help see this practically, we accumulate all gyro rotations into a quaternion, and then display the result as a number.
A user can then observe if rotating the controller from a table surface, up to a vertical surface, registers as exactly 90 degrees in pitch.
Additionally, gyroscopes all tend to drift, and this drift factor regularly changes: there is also an automatic drift compensation system for the instrumentation to make the "real world sensitivity" checks more accurate.
The controller must be kept stationary, and very still, to calibrate the drift factor.
Once an initial check is made, the accumulated rotation is displayed.
The user can also click the "Reset Orientation" button to reset this value.