getting close
but on ps5 i seem to be receiving only 250hz worth of packets while the time stamps are suggesting that we're running at 500hz. I wonder what that's about. Expect it's a bug. Maybe it's down to the hid implementation's sensor rate being initialized to 250... what determines how fast we poll?